Typing notification front end.
Send typing notification events when user types in the compose box.
Listen for these events and display a notification.

Sending notifications:
Underscore's throttle is used for the start notifications (trailing set
to false) and debounce is used for stop notifications.
This is so that start notifications are sent every 10 seconds
and the stop notifications is sent 10 seconds after typing stops.

Displaying notifications:
When a typing notification is received, if the current narrow is private
messages and the user is listed in the notification's recipients (and is not the sender),
"Othello is typing..." is displayed just above the compose div. This notification is
removed after 15 seconds. If another notification is received during this period, the
expiration is extended. When a stop notification is received the notification is removed.