Switch subscriptions overlay to hashchange ignore.
This switches the subscription overlay to operate in a way that is
congruent with the new hashchange ignore logic. The gear-menu trigger
is now a link to #subscriptions which fires an `onhashchange` event,
which triggers an ignore of the hash.

The event listener on exit is also added to unignore the hash state
and return it to its original position before the subscriptions page.