Fix Failing Casper Test.
This is an interesting failure. The reason this needed to change was
because in order for the deactivated users list to re-render, the tab
view needed to be destroyed and re rendered. The subscriptions page
being navigated to now no longer changes the state of the page and
therefore wasn’t the correct navigation for destroying the old admin
page.